Consolidated revit.db into a submodule and added delete.py
