+++ pyrevitlib/pyrevit/revit/db/update.py
@@ -10,35 +10,27 @@ from pyrevit.revit import query
 def update_sheet_revisions(revisions, sheets=None, state=True, doc=None):
     doc or HOST_APP.doc
 
-    # get revisions if not set
-    revisions = revisions or query.get_revisions(doc=doc)
+    # make sure revisions is a list
     if not isinstance(revisions, list):
         revisions = [revisions]
-    # get sheets if not available
-    sheets = sheets or query.get_sheets(doc=doc)
 
-    # cloudedsheets = []
     updated_sheets = []
-    for s in sheets:
-        for r in revisions:
-            # skip issued revisions
-            if not r.Issued:
-                # revs = set([x.IntegerValue for x in s.GetAllRevisionIds()])
-                addrevs = set([x.IntegerValue
-                               for x in s.GetAdditionalRevisionIds()])
-                # cloudrevs = revs - addrevs
-                # if r.Id.IntegerValue in cloudrevs:
-                #     cloudedsheets.append(s)
-                #     continue
+    if revisions:
+        # get sheets if not available
+        for s in sheets or query.get_sheets(doc=doc):
+            addrevs = set([x.IntegerValue
+                           for x in s.GetAdditionalRevisionIds()])
+            for r in revisions:
+                # skip issued revisions
+                if not r.Issued:
+                    if state:
+                        addrevs.add(r.Id.IntegerValue)
+                    elif r.Id.IntegerValue in addrevs:
+                        addrevs.remove(r.Id.IntegerValue)
 
-                if state:
-                    addrevs.add(r.Id.IntegerValue)
-                elif r.Id.IntegerValue in addrevs:
-                    addrevs.remove(r.Id.IntegerValue)
-
-                rev_elids = [DB.ElementId(x) for x in addrevs]
-                s.SetAdditionalRevisionIds(List[DB.ElementId](rev_elids))
-                updated_sheets.append(s)
+            rev_elids = [DB.ElementId(x) for x in addrevs]
+            s.SetAdditionalRevisionIds(List[DB.ElementId](rev_elids))
+            updated_sheets.append(s)
 
     return updated_sheets
